No-list-bucket-upgrade fix
There used to be a check for empty bucket name with a non empty
bucket id, but recent change in b2sdk ensured that an exception
is raised in that case or a bucket is added to cache. This led
to a change of behaviour in tests. This brings the original
behaviour back.